requirements.txt
transformers
huggingface_hub
requests
langchain
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
pymupdf
python-docx
chromadb
mcp
fastapi
uvicorn
pydantic
numpy
scikit-learn
PyPDF2
python-dotenv